Solution to Check Greater Than or Equal To

+++ Check Greater Than or Equal To@@67882a731b4392d8fc1577d8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main()  {
-    float a;
-    scanf("%f",&a);
-    float b;
-    scanf("%f",&b);
+    int a;
+    scanf("%d",&a);
+    int b;
+    scanf("%d",&b);
     if (a>=b){
         printf("True");
     }
